% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixture_model.R
\name{fit_comixture}
\alias{fit_comixture}
\title{Fit shared component gaussian mixed model.}
\usage{
fit_comixture(data, input.parameters, rho = 0.5, max.iterations = 40,
  break.parameter = 5, algorithm = "EM", quiet = FALSE)
}
\arguments{
\item{data}{data.frame with two columns, first being numeric read.counts,
and second being charachter segment names. Column names can be whatever.}

\item{input.parameters}{Data frame holding initialisation parameters for the prior distribution parameters.
With numeric columns "scale" and "shape" the parameters for the gamma
prior on precision; "mean" and "nu" for the mean and precision parameters for the normal prior on 
the mean; and a charachter column "component.type" with either "common" or "specific"}

\item{rho}{Numeric value for initial value of rho parameter}

\item{max.iterations}{maximum number of iterations}

\item{break.parameter}{Numeric value, when the difference in log likelihood 
between two iterations is less than this value the iteration will stop}

\item{algorithm}{Charachter specifying which inference algorithm to use.
One of either "EM" (Expectation-Maximisation) or "VB" (Variational-Bayes)}

\item{quiet, }{Logical specifying whether the iteration number should be printed for each iteration}
}
\value{
A list of parameters
}
\description{
\code{fit_comixture} returns parameters for gassians fitted to the data.
}
\examples{
# see vignette
}

