% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.components}
\alias{plot.components}
\title{Plot fitted shared component gaussian mixed model.}
\usage{
plot.components(vals.df, output.parameters, segment.subset = NULL,
  type = NULL)
}
\arguments{
\item{vals.df}{original data, identical to fit.model() input}

\item{output.parameters, }{list of parameters, output of fit.model()}

\item{segment.subset, }{optional, list of segment names to plot, default uses all segments}

\item{type, }{optional, if value is "density", a density plot comparing orignal vs model distribution.
If value is "QQ", a quantile-quantile plot is output.
If left empty the modeled components are plotted (with common and specific in different colours)}
}
\description{
\code{plot.components} Returns histograms of fitted gaussians and original data by dataset.
}
\examples{
plot.components(input.data, output.data, segment.subset = "segment16")
}

