% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_comixture}
\alias{plot_comixture}
\title{Plot fitted shared component gaussian mixed model.}
\usage{
plot_comixture(data, output.parameters, segment.subset = NULL, type = NULL)
}
\arguments{
\item{data}{original data, identical to fit_comixture() input}

\item{output.parameters, }{list of parameters, output of fit_comixture()}

\item{segment.subset, }{optional, list of segment names to plot, default uses all segments}

\item{type, }{optional, if value is "density", a density plot comparing orignal vs model distribution.
If value is "QQ", a quantile-quantile plot is output.
If left empty the modeled components are plotted (with common and specific in different colours)}
}
\description{
\code{plot_comixture} Returns histograms of fitted gaussians and original data by dataset.
}
\examples{
# plot_comixture(input.data, output.data, segment.subset = "segment16")
}

