% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tidy_output}
\alias{tidy_output}
\title{Convert fit_comixture output into tidy data format}
\usage{
tidy_output(output, segment.subset = NULL)
}
\arguments{
\item{output}{the output given by fit_comixture}

\item{segment.subset, }{optional, list of segment names to plot, default uses all segments}
}
\description{
\code{tidy_output} Convert fit_comixture output into tidy data format .e.g for plotting. This function parses learned/output parameters - the output is given as a nested list, but we need a data frame with one line per component (for each segment) to plot with ggplot2
}
\examples{

}

